server/requirements.txt
fastapi==0.115.4        # FastAPI framework for building APIs
uvicorn==0.32.0         # ASGI server for FastAPI
pydantic==2.9.2         # Data validation and parsing for FastAPI models
gunicorn==23.0.0        # HTTP server compatible with FastAPI
httpx==0.27.2.0
authlib==1.3.2.0
cryptography==43.0.3
pyjwt==2.9.0